Add aggressive inline form validation
This feature was added to provide a better experience for the user.
When focusing out of a field, the page will alert the user of any
incorrect inputs and let the user know how to correct their input until
it meets the form's requirements. This will prevent having the user
wait until the form is submitted to find out a field was filled out
incorrectly.