Refactor code for confirm password validation
This was performed to solve an a case where the user correctly fills
the password fields but decides to change the first password field
right after correctly filling out the confirm password field. This also
solves a bug where if the user submits an invalid form twice and there
is a mismatch of passwords, the error for the confirm password field
disappears on the second submit event.